Tip 1: Give nested resource dedicated controller
